TAN-114 added useAnalysisFilterParams hook to share hook extraction logic
