+++ front/app/api/analysis_tags/__mocks__/useAnalysisTags.ts
@@ -6,6 +6,8 @@ export const tagsData: ITagData[] = [
     type: 'tag',
     attributes: {
       name: 'Tag 1',
+      total_input_count: 4,
+      filtered_input_count: 4,
       tag_type: 'custom',
       created_at: '2020-01-01T00:00:00.000Z',
       updated_at: '2020-01-01T00:00:00.000Z',
